+++ yandex/data_source_yandex_function.go
@@ -128,6 +128,20 @@ func dataSourceYandexFunction() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+
+			"connectivity": {
+				Type:     schema.TypeList,
+				MaxItems: 1,
+				Optional: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"network_id": {
+							Type:     schema.TypeString,
+							Required: true,
+						},
+					},
+				},
+			},
 		},
 	}
 }